requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-pptx>=0.6.21
python-docx>=1.1.0

